small refactor to add domain building utility and helpers directory
