+++ great_expectations/rule_based_profiler/parameter_builder/metric_multi_batch_parameter_builder.py
@@ -1,10 +1,10 @@
 from typing import Any, Dict, List, Optional, Tuple, Union
 
 from great_expectations.core.batch import Batch, BatchRequest, RuntimeBatchRequest
-from great_expectations.rule_based_profiler.types import Domain, ParameterContainer
-from great_expectations.rule_based_profiler.util import (
+from great_expectations.rule_based_profiler.helpers.util import (
     get_parameter_value_and_validate_return_type,
 )
+from great_expectations.rule_based_profiler.types import Domain, ParameterContainer
 
 from great_expectations.rule_based_profiler.parameter_builder.parameter_builder import (  # isort:skip
     MetricComputationResult,
