+++ great_expectations/rule_based_profiler/parameter_builder/simple_date_format_string_parameter_builder.py
@@ -2,6 +2,9 @@ import logging
 from typing import Any, Dict, Iterable, List, Optional, Set, Tuple, Union
 
 from great_expectations.core.batch import Batch, BatchRequest, RuntimeBatchRequest
+from great_expectations.rule_based_profiler.helpers.util import (
+    get_parameter_value_and_validate_return_type,
+)
 from great_expectations.rule_based_profiler.parameter_builder.parameter_builder import (
     AttributedResolvedMetrics,
     MetricComputationResult,
@@ -9,9 +12,6 @@ from great_expectations.rule_based_profiler.parameter_builder.parameter_builder 
     ParameterBuilder,
 )
 from great_expectations.rule_based_profiler.types import Domain, ParameterContainer
-from great_expectations.rule_based_profiler.util import (
-    get_parameter_value_and_validate_return_type,
-)
 
 logger = logging.getLogger(__name__)
